requirements.txt
fastapi
uvicorn
python-dotenv
requests
playwright
pymongo